Changed "Transition.srv" to "SendCommand.srv". Added stubs for InvokeStateChange.srv and SendEvent.srv. Changed command enumeration to match omac guidelines. Added public function to allow event processing for state machine. Removed auto signalling of state_complete_event from runStateMethod.

+++ packml_sm/include/packml_sm/packml_stats_itemized.h
@@ -12,7 +12,7 @@ namespace packml_sm
 struct PackmlStatsItemized
 {
   int16_t id;
-  int32_t count;
+  double count;
   double duration;
 };
 }  // namespace packml_sm
